+++ app/src/pages/Dashboard/CreateEventsPage/CreateEventsPage.tsx
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
 //
 // SPDX-License-Identifier: EUPL-1.2
-import { Box, Step, StepLabel, Stepper, styled } from '@mui/material';
+import { Box, Step, StepLabel, Stepper, styled, Typography } from '@mui/material';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
@@ -10,7 +10,7 @@ import CreateOrUpdateMeetingForm from '../../../components/CreateOrUpdateMeeting
 const Container = styled(Box)(({ theme }) => ({
   display: 'grid',
   gridTemplateColumns: '1fr',
-  gridTemplateRows: 'auto 1fr',
+  gridTemplateRows: 'auto auto 1fr',
   height: '100%',
   overflow: 'hidden auto',
   gap: theme.spacing(2),
@@ -40,6 +40,7 @@ const CreateEventsPage = () => {
 
   return (
     <Container>
+      <Typography component="h1">{t('dashboard-meetings-create-title')}</Typography>
       <StepperHeader />
       {activeStep === 0 && <CreateOrUpdateMeetingForm onForwardButtonClick={() => setActiveStep(1)} />}
     </Container>